Add: 기존의 InputView를 UserInputView로 변경하고, 새로운 InputView 구조체 생성
